+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-25 22:03:07
-// Commit: fix(core): fix memory leak by cancelling core subscription
+// Updated at: 2026-04-26 10:20:20
+// Commit: feat(core): implement shimmer loading skeleton in core
 
 class coreComponentState {
-  final String commitRef = '692';
-  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-25T22:03:07+05:00');
+  final String commitRef = '700';
+  final String lastMessage = 'feat(core): implement shimmer loading skeleton in core';
+  final DateTime timestamp = DateTime.parse('2026-04-26T10:20:20+05:00');
 }