+++ CHANGELOG.md
@@ -701,3 +701,4 @@
 - [2026-04-26 09:46] fix(test): resolve race condition during test initialization
 - [2026-04-26 10:02] fix(test): resolve race condition during test initialization
 - [2026-04-26 10:07] feat(network): add bi-directional event bus handler for network
+- [2026-04-26 10:20] feat(core): implement shimmer loading skeleton in core